+++ ReportStats/src/main/java/dpa/model/AdSetStatsLoader.java
@@ -1,5 +1,6 @@
 package dpa.model;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -7,8 +8,8 @@ import java.util.Date;
  */
 public class AdSetStatsLoader {
 
-    private int Client_ID;
-    private int AdSet_ID;
+    private long Client_ID;
+    private long AdSet_ID;
     private String Product_ID;
     private int Age_Start_Range;
     private int Age_End_Range;
@@ -16,15 +17,15 @@ public class AdSetStatsLoader {
     private String Gender;
     private String Placement;
     private String Impression_Device;
-    private int Reach;
+    private long Reach;
     private double Frequency;
-    private int Clicks;
-    private int Total_Actions;
-    private int Impressions;
-    private int Social_Reach;
-    private int Social_Impressions;
-    private int Unique_Impressions;
-    private int Unique_Social_Impressions;
+    private long Clicks;
+    private long Total_Actions;
+    private long Impressions;
+    private long Social_Reach;
+    private long Social_Impressions;
+    private long Unique_Impressions;
+    private long Unique_Social_Impressions;
     private double CPM;
     private double CPP;
     private double Spend;
@@ -39,19 +40,19 @@ public class AdSetStatsLoader {
     getters and setters
      */
 
-    public int getClient_ID() {
+    public long getClient_ID() {
         return Client_ID;
     }
 
-    public void setClient_ID(int client_ID) {
+    public void setClient_ID(long client_ID) {
         Client_ID = client_ID;
     }
 
-    public int getAdSet_ID() {
+    public long getAdSet_ID() {
         return AdSet_ID;
     }
 
-    public void setAdSet_ID(int adSet_ID) {
+    public void setAdSet_ID(long adSet_ID) {
         AdSet_ID = adSet_ID;
     }
 
@@ -111,11 +112,11 @@ public class AdSetStatsLoader {
         Impression_Device = impression_Device;
     }
 
-    public int getReach() {
+    public long getReach() {
         return Reach;
     }
 
-    public void setReach(int reach) {
+    public void setReach(long reach) {
         Reach = reach;
     }
 
@@ -127,59 +128,59 @@ public class AdSetStatsLoader {
         Frequency = frequency;
     }
 
-    public int getClicks() {
+    public long getClicks() {
         return Clicks;
     }
 
-    public void setClicks(int clicks) {
+    public void setClicks(long clicks) {
         Clicks = clicks;
     }
 
-    public int getTotal_Actions() {
+    public long getTotal_Actions() {
         return Total_Actions;
     }
 
-    public void setTotal_Actions(int total_Actions) {
+    public void setTotal_Actions(long total_Actions) {
         Total_Actions = total_Actions;
     }
 
-    public int getImpressions() {
+    public long getImpressions() {
         return Impressions;
     }
 
-    public void setImpressions(int impressions) {
+    public void setImpressions(long impressions) {
         Impressions = impressions;
     }
 
-    public int getSocial_Reach() {
+    public long getSocial_Reach() {
         return Social_Reach;
     }
 
-    public void setSocial_Reach(int social_Reach) {
+    public void setSocial_Reach(long social_Reach) {
         Social_Reach = social_Reach;
     }
 
-    public int getSocial_Impressions() {
+    public long getSocial_Impressions() {
         return Social_Impressions;
     }
 
-    public void setSocial_Impressions(int social_Impressions) {
+    public void setSocial_Impressions(long social_Impressions) {
         Social_Impressions = social_Impressions;
     }
 
-    public int getUnique_Impressions() {
+    public long getUnique_Impressions() {
         return Unique_Impressions;
     }
 
-    public void setUnique_Impressions(int unique_Impressions) {
+    public void setUnique_Impressions(long unique_Impressions) {
         Unique_Impressions = unique_Impressions;
     }
 
-    public int getUnique_Social_Impressions() {
+    public long getUnique_Social_Impressions() {
         return Unique_Social_Impressions;
     }
 
-    public void setUnique_Social_Impressions(int unique_Social_Impressions) {
+    public void setUnique_Social_Impressions(long unique_Social_Impressions) {
         Unique_Social_Impressions = unique_Social_Impressions;
     }
